refactor: extract manager logic into hooks
