+++ components/WorkScheduleManager/useWorkScheduleManager.ts
@@ -1,0 +1,7 @@
+import { usePlanningReminderNotifications } from './usePlanningReminderNotifications';
+import { useWorkScheduleSuggestion } from './useWorkScheduleSuggestion';
+
+export function useWorkScheduleManager() {
+  useWorkScheduleSuggestion();
+  usePlanningReminderNotifications();
+}
